Clean up code per clippy best practices and harden token file permissions
- Fix all clippy warnings across the workspace (io::Error::other,
  redundant Ok(...?), unnecessary casts, div_ceil, while-let loop,
  unused assignment via loop-break value)
- Remove dead fields from InodeEntry/SymlinkEntry and drop the now-unused
  get_or_create_inode parameters
- Write tokens.json with 0600 permissions so OAuth tokens are not
  world-readable

+++ crates/sync-engine/src/engine.rs
@@ -985,7 +985,7 @@ impl SyncEngine {
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-                let free = stat.f_bavail as u64 * stat.f_bsize as u64;
+                let free = stat.f_bavail * stat.f_bsize;
                 if free < Self::MIN_FREE_BYTES {
                     let free_mb = free / (1024 * 1024);
                     anyhow::bail!(
